+++ src/retaggr/engines/__init__.py
@@ -2,7 +2,6 @@ from .base import Engine
 from .base import ImageResult
 
 from .danbooru import Danbooru
-from .e621 import E621
 from .iqdb import Iqdb
 from .paheal import Paheal
 from .saucenao import SauceNao