+++ src/retaggr/engines/saucenao/engine.py
@@ -3,6 +3,7 @@ import asyncio
 import functools
 
 from retaggr.engines.base import Engine, ImageResult
+from retaggr.engines.saucenao.handlers import DanbooruHandler, E621Handler, KonachanHandler, YandereHandler
 from retaggr.errors import NotAvailableSearchException, EngineCooldownException
 import requests as fuck_aiohttp
 
@@ -15,12 +16,13 @@ class SauceNao(Engine):
 
     :param api_key: SauceNao API key. You can get this by registering an account on saucenao.com
     :type api_key: str
+    :param test_mode: Enable test mode. Test mode is unique in that it does not need an API key, but it only works on one URL.
     """
     host = "https://saucenao.com"
     download_required = False
 
 
-    tag_indexes = set([9, 12, 26])
+    tag_indexes = set([9, 12, 26, 29])
     """Tag indexes we can parse.
 
     Valid index numbers can be found at https://saucenao.com/status.html .
@@ -30,9 +32,10 @@ class SauceNao(Engine):
     * 9: Danbooru
     * 12: Yande.re
     * 26: Konachan
+    * 29: E621
     """
 
-    source_indexes = [5, 16, 37, 34]
+    source_indexes = [5, 16, 29, 37, 34]
     """List of source indexes in preferred order (key 0 is preferred, last key is least preferred).
     
     Valid index numbers can be found at https://saucenao.com/status.html .
@@ -44,8 +47,27 @@ class SauceNao(Engine):
     * 37: MangaDex (not official redistribution, but metadata is accurate)
     * 34: DeviantART (not preferred, large number of art theft and reuploads)
     """
-    def __init__(self, api_key):
+
+    indirect_source_indexes = [29]
+    """List of indirect source indexes.
+
+    There is no preferred order.
+
+    Some of these may require extra steps to obtain the source.
+
+    * 29: E621
+    """
+    def __init__(self, api_key, test_mode=False):
         self.api_key = api_key
+        self.handlers = {
+            DanbooruHandler.engine_id : DanbooruHandler(),
+            KonachanHandler.engine_id : KonachanHandler(),
+            YandereHandler.engine_id : YandereHandler(),
+        }
+
+    def enable_e621(self, username, app_name, version):
+        """Activate the E621 parser."""
+        self.handlers[E621Handler.engine_id] = E621Handler(username, app_name, version)
 
     async def search_image(self, url):
         request_url = "https://saucenao.com/search.php"
@@ -72,45 +94,26 @@ class SauceNao(Engine):
         :param json: JSON output from the API.
         :type json: dict
         :return: Dictionary containing data that matches the output for :meth:`SauceNao.search_image_source`
-        :rtype: dict
+        :rtype: ImageResult
         """
         base_similarity = json["header"]["minimum_similarity"] # Grab the minimum similarity saucenao advises, going lower is generally gonna give false positives.
 
         # Below we cast the _entry_ similarity to a float since somehow it's stored as an str.
         # Damn API inaccuracy
-        source_results = [entry for entry in json["results"] if entry["header"]["index_id"] in self.source_indexes and float(entry["header"]["similarity"]) > base_similarity]
-
-        source = None
-        source_priority = len(self.source_indexes) # No result priority is 1 above the least wanted result
-        for entry in source_results:
-            list_index = self.source_indexes.index(entry["header"]["index_id"])
-            if list_index < source_priority: # If our exsting result priority is lower than the current result...
-                source = entry["data"]["ext_urls"][0] # We update the source with that result
-                source_priority = list_index # And we update the priority
-
-        # See my comment above source_results as to what I'm doing here.
-        tag_results_list = [entry for entry in json["results"] if entry["header"]["index_id"] in self.tag_indexes and float(entry["header"]["similarity"]) > base_similarity]
-
-        # Unlike the source, these require specific lookups based on their ID. As a result, I'll rearrange the results to a dict.
-        tag_results = {}
-        for entry in tag_results_list:
-            tag_results[entry["header"]["index_id"]] = entry
+        valid_results = [entry for entry in json["results"] if float(entry["header"]["similarity"]) > base_similarity]
 
         # Kinda looks stupid, but whatever.
         loop = asyncio.get_event_loop()
-        tags = []
-        for index_id, entry in tag_results.items():
-            if index_id == 9: # Danbooru
-                r = await loop.run_in_executor(None, functools.partial(fuck_aiohttp.get, "https://danbooru.donmai.us/posts/" + str(entry["data"]["danbooru_id"]) + ".json"))
-                j = r.json()
-                tags += j["tag_string"].split()
-            if index_id == 12: # Yande.re # pragma: no cover
-                r = await loop.run_in_executor(None, functools.partial(fuck_aiohttp.get, "https://yande.re/post.json", params={"tags": "id:" + str(entry["data"]["yandere_id"])}))
-                j = r.json()
-                tags += j[0]["tags"].split()
-            if index_id == 26: # Konachan # pragma: no cover
-                r = await loop.run_in_executor(None, functools.partial(fuck_aiohttp.get, "http://konachan.com/post.json", params={"tags": "id:" + str(entry["data"]["konachan_id"])}))
-                j = r.json()
-                tags += j[0]["tags"].split()
+        source = set()
+        tags = set()
+        for entry in valid_results:
+            for url in entry["data"]["ext_urls"]:
+                source.add(url)
+            handler = self.handlers.get(entry["header"]["index_id"], None)
+            if handler:
+                if handler.tag_capable:
+                    tags.update(await handler.get_tag_data(entry["data"]))
+                if handler.source_capable:
+                    source.update(await handler.get_source_data(entry["data"]))
 
         return ImageResult(tags, source, None)